+++ Common/src/main/java/edu/sdccd/cisc191/template/GroceryRequest.java
@@ -1,8 +1,12 @@
 package edu.sdccd.cisc191.template;
 import java.io.Serializable;
 public class GroceryRequest implements Serializable {
-    public String category;
-    public String name;
+
+    //Sean Standen - Peer Review
+    //Made instance variables private so they can only
+    //be accessed by getter and setter methods.
+    private String category;
+    private String name;
     public GroceryRequest(String category, String name) {
         this.category = category;
         this.name = name;
