+++ Common/src/main/resources/Prices.csv
@@ -10,3 +10,11 @@ Green Onion, $0.89, Organic
 Yellow Onion, $1.43, Organic
 Yellow Onion, $1.03, Not Organic
 Blood Orange, $2.99, Not Organic
+Zucchini Squash, $0.77, Not Organic
+Baby Bella Mushroom, $2.99, Organic
+Zucchini Squash, $0.77, Not Organic
+Baby Bella Mushroom, $2.99, Organic
+Zucchini Squash, $0.77, Not Organic
+Baby Bella Mushroom, $2.99, Organic
+Zucchini Squash, $0.77, Not Organic
+Baby Bella Mushroom, $2.99, Organic
